aura-ml/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-dotenv>=1.0.0
supabase>=1.0.0
sentence-transformers>=2.2.0
torch>=2.0.0
scikit-learn>=1.3.0
pandas>=2.0.0
numpy>=1.24.0
librosa>=0.10.0
transformers>=4.30.0
textblob>=0.17.0
Pillow>=10.0.0
requests>=2.31.0
python-multipart>=0.0.6
pytest>=7.0.0
